+++ CHANGELOG.md
@@ -4,6 +4,9 @@
  *   Allow breakpoints and stepping to work together in debug
  *   API_uploadCode() allows to immediately run the code: Regarding preload it defined whether the code should be preloaded in the whiteboard or run when the user clicks run
  *   Breakpoints in debug now count how many times the app has gone through them (most useful when the breakpoint exists but is disabled)
+ *   Build view now also shows visual setup, although by default it shows the advanced setup
+ *   Added tips, initial, minValue, etc to all parameters in instructions
+ *   Added a few new parameters to existing instructions
 
 ## 2.1 (2015-10-06)